==> Supply and demand curves on Nord Pool/1 Download data.ipynb <==
import pandas as pd
#!pip install xlrd wget

df = pd.read_excel("https://www.nordpoolgroup.com/globalassets/download-center-market-data/mcp_data_report_20-12-2018-00_00_00.xls")
df.head()

import datetime
def date_range(start_date, end_date):
    delta = datetime.timedelta(days=1)
    d = start_date
    while d <= end_date:
        yield(d)
        d += delta
        
[d.strftime("%Y-%m-%d") for d in date_range(start_date = datetime.date(2009, 5, 25), end_date = datetime.date(2009, 5, 30))]

#!pip install wget
import wget
import time

#start_date = datetime.date(2014, 7, 1)
start_date = datetime.date(2020, 1, 13)
end_date = datetime.date.today()

for date in date_range(start_date, end_date):
    try:
        #https://www.nordpoolgroup.com/globalassets/download-center-market-data/mcp_data_report_20-12-2018-00_00_00.xls
        url = "https://www.nordpoolgroup.com/globalassets/download-center-market-data/mcp_data_report_%s-00_00_00.xls" % date.strftime("%d-%m-%Y")
        file_name = wget.download(url)
        print(file_name)
        time.sleep(0.1)
    except:
        print(url + " FAILED!")
    else:
        print(file_name)
